Add stream in places list
